+++ arki/aws/ecs.py
@@ -2,9 +2,16 @@ import boto3
 import click
 import json
 import logging
-import sys
-from arki.aws.base_helper import BaseHelper
+from os.path import basename
+from arki.configs import (
+    init_wrapper,
+    default_config_file_path,
+)
 
+APP_NAME = basename(__file__).split('.')[0]
+
+# Default configuration file location
+DEFAULT_CONFIG_FILE = default_config_file_path(f"{APP_NAME}.toml")
 
 DEFAULT_CONFIGS = {
     "aws.profile": {"required": True},
@@ -32,31 +39,38 @@ def task_definition(arn):
     print(json.dumps(resp["taskDefinition"], sort_keys=True, indent=2))
 
 
-@click.command()
-@click.argument("ini_file", required=False)
-@click.option("--init", "-i", is_flag=True, help="Set up new configuration")
-@click.option("--family", "-f", required=False, help="Filtered by family name (ake task definition name")
-@click.option("--detail", "-d", required=False, help="ARN of the task definition to show more details")
-def main(ini_file, init, family, detail):
-    """
-    ECS
-
-    Use --init to create a `ini_file` with the default template to start.
-    """
-
+@init_wrapper
+def process(*args, **kwargs):
     try:
-        helper = BaseHelper(DEFAULT_CONFIGS, ini_file)
+        detail = kwargs.get("detail")
+        family = kwargs.get("family")
 
-        if init:
-            helper._create_ini_template(module=__file__, allow_overriding_default=True)
+        if detail:
+            task_definition(arn=detail)
         else:
-            if detail:
-                task_definition(arn=detail)
-            else:
-                list_task_definitions(family_name=family)
+            list_task_definitions(family_name=family)
 
     except Exception as e:
         logging.error(e)
-        sys.exit(1)
+        return 1
 
-    sys.exit(0)
+    return 0
+
+
+@click.command()
+@click.argument("config_file", required=False, default=DEFAULT_CONFIG_FILE)
+@click.option("--config_section", "-s", required=False, default=APP_NAME, help=f"E.g. {APP_NAME}.staging")
+@click.option("--family", "-f", required=False, help="Filtered by family name (ake task definition name")
+@click.option("--detail", "-d", required=False, help="ARN of the task definition to show more details")
+def main(config_file, config_section, family, detail):
+    """
+    ECS
+    """
+    process(
+        app_name=APP_NAME,
+        config_file=config_file,
+        default_configs=DEFAULT_CONFIGS,
+        config_section=config_section,
+        family=family,
+        detail=detail,
+    )