+++ arki/aws/lambda_deploy.py
@@ -1,11 +1,18 @@
 import boto3
 import click
 import logging
-import sys
-from arki.aws import check_response
-from arki.aws.base_helper import BaseHelper
-from arki.configs import tokenize_multiline_values
+from os.path import basename
 
+from arki.aws import check_response
+from arki.configs import (
+    init_wrapper,
+    default_config_file_path,
+)
+
+APP_NAME = basename(__file__).split('.')[0]
+
+# Default configuration file location
+DEFAULT_CONFIG_FILE = default_config_file_path(f"{APP_NAME}.toml")
 
 DEFAULT_CONFIGS = {
     "aws.profile": {"required": False},
@@ -19,12 +26,6 @@ DEFAULT_CONFIGS = {
 }
 
 
-def retrieve_lambda_environments(settings):
-    """Retrieve environment variables to be set for a stage
-    """
-    return tokenize_multiline_values(settings, "aws.lambda.environment")
-
-
 def prepare_func_configuration_args(settings, description):
     """Prepare arguments for calling update_function_configuration()
     """
@@ -35,8 +36,8 @@ def prepare_func_configuration_args(settings, description):
         "Handler": settings["aws.lambda.handler"],
         "Role": settings["aws.lambda.role.arn"],
         "Runtime": settings["aws.lambda.runtime"],
-        "MemorySize": int(settings["aws.lambda.memory"]),
-        "Timeout": int(settings["aws.lambda.timeout"]),
+        "MemorySize": settings["aws.lambda.memory"],
+        "Timeout": settings["aws.lambda.timeout"],
     }
 
     kms_key_arn = settings.get("aws.lambda.kmskey.arn")
@@ -44,7 +45,7 @@ def prepare_func_configuration_args(settings, description):
         func_configuration_args["KMSKeyArn"] = kms_key_arn
 
     # Retrieve lambda environment variables in dict form
-    env_vars = retrieve_lambda_environments(settings)
+    env_vars = settings["aws.lambda.environment"]
     if len(env_vars) > 0:
         func_configuration_args["Environment"] = {"Variables": env_vars}
     return func_configuration_args
@@ -53,10 +54,10 @@ def prepare_func_configuration_args(settings, description):
 def _lambda_deploy(settings, zipfile, description):
 
     lambda_function = settings["aws.lambda.name"]
-    lambda_alias = settings["aws.lambda.alias"]
+    lambda_alias = settings.get("aws.lambda.alias")
     func_configuration_args = prepare_func_configuration_args(settings, description)
 
-    logging.info(f"Deploying to {lambda_alias}: {func_configuration_args}")
+    logging.info(f"Deploying (alias={lambda_alias}): {func_configuration_args}")
 
     with open(zipfile, mode="rb") as fh:
         byte_stream = fh.read()
@@ -80,41 +81,52 @@ def _lambda_deploy(settings, zipfile, description):
     check_response(resp)
     logging.info(f"Updated function config: Version={resp['Version']}, RevisionId={resp['RevisionId']}")
 
-    resp = client.update_alias(
-        FunctionName=lambda_function,
-        Name=lambda_alias,
-        FunctionVersion=version,
-        Description=description,
-    )
-    check_response(resp)
-    logging.info(f"Updated alias on Version={resp['FunctionVersion']}: AliasArn={resp['AliasArn']}")
+    if lambda_alias:
+        resp = client.update_alias(
+            FunctionName=lambda_function,
+            Name=lambda_alias,
+            FunctionVersion=version,
+            Description=description,
+        )
+        check_response(resp)
+        logging.info(f"Updated alias on Version={resp['FunctionVersion']}: AliasArn={resp['AliasArn']}")
 
 
-@click.command()
-@click.argument("ini_file", required=True)
-@click.option("--init", "-i", is_flag=True, help="Set up new configuration")
-@click.option("--zipfile", "-z", required=True, help="Zip file of the lambda function code")
-@click.option("--alias", "-a", required=True, help="Alias of the deployment")
-@click.option("--description", "-d", required=True, help="Description of the deployment")
-def lambda_deploy(ini_file, init, zipfile, alias, description):
-    """
-    Use --init to create a `ini_file` with the default template to start.
-    """
+@init_wrapper
+def process(*args, **kwargs):
+    logging.debug("At process")
+    logging.debug(kwargs)
 
     try:
-        helper = BaseHelper(DEFAULT_CONFIGS, ini_file, stage_section=alias)
+        settings = kwargs.get("_arki_settings")
+        zipfile = kwargs.get("zipfile")
+        description = kwargs.get("description")
 
-        if init:
-            helper._create_ini_template(module=__file__, allow_overriding_default=True)
-        else:
-            _lambda_deploy(
-                settings=helper.settings,
-                zipfile=zipfile,
-                description=description
-            )
+        _lambda_deploy(
+            settings=settings,
+            zipfile=zipfile,
+            description=description
+        )
 
     except Exception as e:
         logging.error(e)
-        sys.exit(1)
+        return 1
 
-    sys.exit(0)
+    return 0
+
+
+@click.command()
+@click.argument("config_file", required=False, default=DEFAULT_CONFIG_FILE)
+@click.option("--config_section", "-s", required=False, default=APP_NAME, help=f"E.g. {APP_NAME}.staging")
+@click.option("--zipfile", "-z", required=True, help="Zip file of the lambda function code")
+@click.option("--description", "-d", required=True, help="Description of the deployment")
+def lambda_deploy(config_file, config_section, zipfile, description):
+
+    process(
+        app_name=APP_NAME,
+        config_file=config_file,
+        default_configs=DEFAULT_CONFIGS,
+        config_section=config_section,
+        zipfile=zipfile,
+        description=description,
+    )